+++ terraform/locals.tf
@@ -33,4 +33,8 @@ locals {
     dp_file_csi_driver       = "${var.cluster_name}-dp-file-csi-driver"
     dp_image_registry        = "${var.cluster_name}-dp-image-registry"
   }
+
+  hcp_api_version             = "2026-06-30-preview"
+  managed_resource_group_name = coalesce(var.managed_resource_group_name, "${var.cluster_name}-managed")
+  etcd_encryption_key_name    = "etcd-data-kms-encryption-key"
 }